refactor: move initialization logic to the state
